d-ai-app-patterns/scalable_chat/src/sse_service/pyproject.toml
[project]
name = "scalable-chat-sse-service"
version = "0.1.0"
description = "SSE streaming service for scalable chat application"
authors = [
    {name = "Developer"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "azure-servicebus>=7.12.0",
    "azure-identity>=1.15.0",
    "azure-monitor-opentelemetry>=1.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "aiohttp>=3.12.6",
]
requires-python = ">=3.11"
